+++ higher-order-effects.cabal
@@ -13,8 +13,7 @@ extra-source-files:  ChangeLog.md
 cabal-version:       >=1.10
 
 library
-  exposed-modules:     Control.Carrier
-                     , Control.Effect
+  exposed-modules:     Control.Effect
                      , Control.Effect.Cut
                      , Control.Effect.Error
                      , Control.Effect.Fail
